restore videoapi test
